% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABSurvTDC.R
\name{ABCoxPH}
\alias{ABCoxPH}
\title{Cox-PH Model for Animal Breeding}
\usage{
ABCoxPH(wide_data, lact)
}
\arguments{
\item{wide_data}{Dataset from DataPrep function}

\item{lact}{Number of lactation to be used for model building}
}
\value{
\itemize{
  \item Cox_Model - ABCoxPH model
  \item LongData- Long data
}
}
\description{
Data preparation for ABCoxPH
}
\examples{
library("ABSurvTDC")
library("readxl")
data_test<-read_excel(path = system.file("extdata/data_test.xlsx", package = "ABSurvTDC"))
PropData<-DataPrep(data =as.data.frame(data_test))
ABCoxPH(PropData)
}
\references{
\itemize{
\item J.D. Kalbfleisch and R.L. Prentice (1980). The statistical analysis of failure time data. John Wiley & Sons, Inc., New York, 1980.  <doi:10.1002/9781118032985>
\item J.P. Klein and  M L. Moeschberger (2003). Survival Analysis: Techniques for Censored and Truncated Data. Springer New York. <doi:10.1007/b97377>
}
}
